requirements.txt
cmake
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.9.2
Pillow==7.0.0
face_recognition
requests

